refactor: extract generic web HTTP routes (#1734)
