+++ control_plane/http_app.py
@@ -44,13 +44,16 @@ from control_plane.http_routes import (
     DriverReadRouteDependencies,
     EVIDENCE_INGRESS_ROUTES as _EVIDENCE_INGRESS_ROUTES,
     EvidenceWriteRouteDependencies,
+    GenericWebWriteRouteDependencies,
     ProductReadRouteDependencies,
     PromotionEvidenceRequest as PromotionEvidenceRequest,
     ReadRouteDependencies,
     WorkGraphReadRouteDependencies,
     accepted_evidence_response,
+    build_generic_web_write_route_handlers,
     idempotency_capable_store,
     idempotency_scope as idempotency_scope,
+    provider_operation_response_payload as _provider_operation_response_payload,
     product_profile_context_cutover_contexts_allowed,
     register_agent_context_read_routes,
     register_deployment_promotion_read_routes,
@@ -61,6 +64,8 @@ from control_plane.http_routes import (
     register_every_code_notification_attempt_read_routes,
     register_every_code_preview_gate_read_routes,
     register_every_code_work_request_read_routes,
+    register_generic_web_rollback_write_routes,
+    register_generic_web_write_routes,
     register_ingress_read_routes,
     register_inventory_operation_read_routes,
     register_managed_secret_read_routes,
@@ -191,61 +196,9 @@ from control_plane.launchplane_self_deploy_http import (
     LaunchplaneSelfDeployEnvelope,
     launchplane_self_deploy_records,
 )
-from control_plane.generic_web_rollback_http import (
-    GENERIC_WEB_ROLLBACK_PLAN_ROUTE as _GENERIC_WEB_ROLLBACK_PLAN_ROUTE,
-    GENERIC_WEB_ROLLBACK_ROUTE as _GENERIC_WEB_ROLLBACK_ROUTE,
-    GenericWebRollbackEnvelope,
-    GenericWebRollbackPlanEnvelope,
-    GenericWebRollbackProductMismatchError,
-    GenericWebRollbackRouteDependencyError,
-    execute_generic_web_rollback_plan_result,
-    execute_generic_web_rollback_result,
-    resolve_generic_web_rollback_lane,
-    should_store_generic_web_rollback_idempotency,
-)
-from control_plane.generic_web_deploy_http import (
-    GENERIC_WEB_DEPLOY_ROUTE as _GENERIC_WEB_DEPLOY_ROUTE,
-    GenericWebDeployEnvelope,
-    GenericWebDeployProductMismatchError,
-    GenericWebDeployRouteDependencyError,
-    execute_generic_web_deploy_result,
-    resolve_generic_web_deploy_lane,
-    should_store_generic_web_deploy_idempotency,
-)
-from control_plane.workflows.generic_web_deploy import (
-    GenericWebDeployStore,
-    normalize_generic_web_artifact_id,
-    record_observed_generic_web_deploy,
-)
-from control_plane.workflows.generic_web_deploy_provider import (
-    GenericWebDeployProvider,
-    GenericWebResolvedDeployTarget,
-    build_generic_web_provider_reconciliation_key,
-    build_generic_web_provider_target_key,
-    default_generic_web_deploy_provider,
-    generic_web_provider_deployment_succeeded,
-    resolve_generic_web_provider_reconciliation_target,
-)
-from control_plane.workflows.odoo_generic_web_post_deploy import (
-    generic_web_post_deploy_executor_for_driver_id,
-)
 from control_plane.generic_web_promotion_http import (
-    GENERIC_WEB_PROD_PROMOTION_ROUTE as _GENERIC_WEB_PROD_PROMOTION_ROUTE,
-    GENERIC_WEB_PROD_PROMOTION_WORKFLOW_ROUTE as _GENERIC_WEB_PROD_PROMOTION_WORKFLOW_ROUTE,
-    GenericWebProdPromotionEnvelope,
-    GenericWebProdPromotionResponse,
-    GenericWebProdPromotionResponseResult,
-    GenericWebPromotionProductMismatchError,
-    GenericWebPromotionRouteDependencyError,
-    GenericWebPromotionWorkflowEnvelope,
-    GenericWebPromotionWorkflowResponse,
-    GenericWebPromotionWorkflowResponseResult,
     dispatch_generic_web_promotion_workflow_result,
     execute_generic_web_prod_promotion_result,
-    resolve_generic_web_promotion_destination_lane,
-    resolve_generic_web_promotion_workflow_lane,
-    should_store_generic_web_promotion_idempotency,
-    validate_generic_web_prod_promotion_lanes,
 )
 from control_plane.product_promotion_http import (
     PRODUCT_PROMOTION_DRY_RUN_MARKER_ROUTE as _PRODUCT_PROMOTION_DRY_RUN_MARKER_ROUTE,
@@ -263,24 +216,9 @@ from control_plane.product_promotion_http import (
     product_promotion_continuity_payload,
     product_promotion_direct_request,
     product_promotion_dry_run_key,
-    product_promotion_intent_matches,
     product_promotion_request_payload,
     product_promotion_workflow_request,
     resolve_product_promotion_target,
-)
-from control_plane.generic_web_verification_http import (
-    GENERIC_WEB_PREVIEW_VERIFICATION_ROUTE as _GENERIC_WEB_PREVIEW_VERIFICATION_ROUTE,
-    GENERIC_WEB_STABLE_VERIFICATION_ROUTE as _GENERIC_WEB_STABLE_VERIFICATION_ROUTE,
-    GenericWebPreviewVerificationEnvelope,
-    GenericWebStableVerificationEnvelope,
-    GenericWebVerificationProductMismatchError,
-    GenericWebVerificationRouteDependencyError,
-    apply_generic_web_preview_verification_result,
-    apply_generic_web_stable_verification_result,
-    generic_web_verification_response_records,
-    resolve_generic_web_preview_verification_profile,
-    resolve_generic_web_stable_verification_lane,
-    should_store_generic_web_verification_idempotency,
 )
 from control_plane.verireel_read_http import (
     VERIREEL_PREVIEW_DESTROY_ROUTE as _VERIREEL_PREVIEW_DESTROY_ROUTE,
@@ -334,24 +272,6 @@ from control_plane.verireel_prod_http import (
     apply_verireel_prod_promotion_result,
     apply_verireel_prod_rollback_result,
     should_store_verireel_prod_result_idempotency,
-)
-from control_plane.generic_web_preview_http import (
-    GENERIC_WEB_PREVIEW_DESTROY_ROUTE as _GENERIC_WEB_PREVIEW_DESTROY_ROUTE,
-    GENERIC_WEB_PREVIEW_INVENTORY_ROUTE as _GENERIC_WEB_PREVIEW_INVENTORY_ROUTE,
-    GENERIC_WEB_PREVIEW_READINESS_ROUTE as _GENERIC_WEB_PREVIEW_READINESS_ROUTE,
-    GENERIC_WEB_PREVIEW_REFRESH_ROUTE as _GENERIC_WEB_PREVIEW_REFRESH_ROUTE,
-    GenericWebPreviewDestroyEnvelope,
-    GenericWebPreviewInventoryEnvelope,
-    GenericWebPreviewProductMismatchError,
-    GenericWebPreviewReadinessEnvelope,
-    GenericWebPreviewRefreshEnvelope,
-    GenericWebPreviewRouteDependencyError,
-    apply_generic_web_preview_destroy_result,
-    apply_generic_web_preview_inventory_result,
-    apply_generic_web_preview_readiness_result,
-    apply_generic_web_preview_refresh_result,
-    resolve_generic_web_preview_profile,
-    should_store_generic_web_preview_idempotency,
 )
 from control_plane.odoo_artifact_publish_inputs_http import (
     ODOO_ARTIFACT_PUBLISH_INPUTS_ROUTE as _ODOO_ARTIFACT_PUBLISH_INPUTS_ROUTE,
@@ -510,10 +430,6 @@ from control_plane.drivers import native_routes
 from control_plane.drivers.route_paths import (
     INGRESS_ROUTE_APPLY_ROUTE as _INGRESS_ROUTE_APPLY_ROUTE,
 )
-from control_plane.drivers.generic_web_preview_dispatch import (
-    GenericWebPreviewDesiredStateEnvelope,
-    _GENERIC_WEB_PREVIEW_DESIRED_STATE_ROUTE,
-)
 from control_plane.every_code_work_request_write import (
     EveryCodeWorkRequestCreateEnvelope,
     build_every_code_work_request_record,
@@ -629,11 +545,6 @@ from control_plane.workflows.preview_pr_feedback import (
     EveryCodeWorkRequestReadStore,
     PreviewPrFeedbackPreviewReadStore,
     build_preview_pr_feedback_record,
-)
-from control_plane.workflows.generic_web_deploy import product_profile_uses_generic_web_base
-from control_plane.workflows.generic_web_preview import (
-    GenericWebPreviewProfileStore,
-    discover_generic_web_preview_desired_state,
 )
 from control_plane.workflows.launchplane_self_deploy import execute_launchplane_self_deploy
 from control_plane.workflows.ship import utc_now_timestamp
@@ -1313,19 +1224,6 @@ class EveryCodePrFeedbackStatusEnvelope(BaseModel):
         return self
 
 
-def _provider_operation_response_payload(
-    *,
-    trace_id: str,
-    records: Mapping[str, object],
-    result: dict[str, object],
-) -> dict[str, object]:
-    return AcceptedEvidenceResponse(
-        trace_id=trace_id,
-        records=dict(records),
-        result=result,
-    ).model_dump(mode="json", exclude_none=True)
-
-
 class _OdooPreviewProviderMutationAdapter:
     def __init__(
         self,
@@ -1423,276 +1321,6 @@ class _OdooPreviewProviderMutationAdapter:
             ),
             durable=not driver_result_contains_status(driver_result, "blocked")
             and driver_status != "fail",
-            provider_effect_performed=provider_effect_attempted,
-        )
-
-
-class _GenericWebDeployProviderMutationAdapter:
-    def __init__(
-        self,
-        *,
-        control_plane_root: FilePath,
-        record_store: object,
-        deploy_request: GenericWebDeployEnvelope,
-        profile: LaunchplaneProductProfileRecord,
-        lane: ProductLaneProfile,
-        trace_id: str,
-    ) -> None:
-        self._control_plane_root = control_plane_root
-        self._record_store = record_store
-        self._deploy_request = deploy_request
-        self._profile = profile
-        self._lane = lane
-        self._trace_id = trace_id
-        self._deploy_provider: GenericWebDeployProvider = default_generic_web_deploy_provider()
-        self._resolved_deploy_target: GenericWebResolvedDeployTarget | None = None
-
-    def _resolve_deploy_target(self) -> GenericWebResolvedDeployTarget:
-        if self._resolved_deploy_target is None:
-            deploy = self._deploy_request.deploy
-            self._resolved_deploy_target = self._deploy_provider.resolve_deploy_target(
-                control_plane_root=self._control_plane_root,
-                request_artifact_id=deploy.artifact_id,
-                request_source_git_ref=deploy.source_git_ref,
-                request_timeout_seconds=deploy.timeout_seconds,
-                request_no_cache=deploy.no_cache,
-                record_store=self._record_store,
-                profile=self._profile,
-                lane=self._lane,
-                normalized_artifact_id=normalize_generic_web_artifact_id(
-                    profile=self._profile,
-                    artifact_id=deploy.artifact_id,
-                ),
-                fallback_target_name=f"{self._profile.product}-{self._lane.instance}",
-            )
-        return self._resolved_deploy_target
-
-    def reconciliation_key(self) -> str:
-        return build_generic_web_provider_reconciliation_key(self._resolve_deploy_target())
-
-    def target_key(self) -> str:
-        return build_generic_web_provider_target_key(self._resolve_deploy_target())
-
-    def observe(
-        self,
-        provider_operation_key: str,
-        provider_effect_phase: str,
-        reconciliation_key: str,
-    ) -> ProviderObservation:
-        try:
-            deploy = self._deploy_request.deploy
-            resolved_target = resolve_generic_web_provider_reconciliation_target(
-                reconciliation_key=reconciliation_key,
-                request_artifact_id=deploy.artifact_id,
-                request_source_git_ref=deploy.source_git_ref,
-                request_timeout_seconds=deploy.timeout_seconds,
-                request_no_cache=deploy.no_cache,
-                normalized_artifact_id=normalize_generic_web_artifact_id(
-                    profile=self._profile,
-                    artifact_id=deploy.artifact_id,
-                ),
-                lane=self._lane,
-            )
-            self._resolved_deploy_target = resolved_target
-            observation = self._deploy_provider.observe_artifact_deploy(
-                control_plane_root=self._control_plane_root,
-                resolved_deploy_target=resolved_target,
-                deployment_title=provider_operation_title(provider_operation_key),
-            )
-        except (FileNotFoundError, ValueError, click.ClickException):
-            return ProviderObservation(outcome="unknown")
-        if observation.outcome != "present":
-            if observation.outcome == "absent" and provider_effect_phase == "deploy_trigger":
-                return ProviderObservation(outcome="unknown")
-            return ProviderObservation(
-                outcome=observation.outcome,
-                retry_safe=(
-                    observation.outcome == "absent"
-                    and provider_effect_phase in {"", "target_update"}
-                ),
-            )
-        post_deploy_unobserved = (
-            generic_web_provider_deployment_succeeded(observation.deployment_status)
-            and generic_web_post_deploy_executor_for_driver_id(self._profile.driver_id) is not None
-        )
-        deployment_record_id = self._deployment_record_id(provider_operation_key)
-        if provider_effect_phase.startswith("post_deploy_"):
-            read_deployment_record = getattr(self._record_store, "read_deployment_record", None)
-            if not callable(read_deployment_record):
-                return ProviderObservation(outcome="unknown")
-            try:
-                read_deployment_record(deployment_record_id)
-            except FileNotFoundError:
-                return ProviderObservation(outcome="unknown")
-            post_deploy_unobserved = False
-        try:
-            records, driver_result = record_observed_generic_web_deploy(
-                record_store=cast(GenericWebDeployStore, self._record_store),
-                profile=self._profile,
-                lane=self._lane,
-                resolved_deploy_target=resolved_target,
-                observation=observation,
-                deployment_record_id=deployment_record_id,
-                post_deploy_unobserved=post_deploy_unobserved,
-            )
-        except (FileNotFoundError, ValueError, click.ClickException):
-            return ProviderObservation(outcome="unknown")
-        terminal_failure = str(driver_result.get("deploy_status", "")).strip() == "fail"
-        return ProviderObservation(
-            outcome="present",
-            response_status_code=502 if terminal_failure else 202,
-            response_payload=_provider_operation_response_payload(
-                trace_id=self._trace_id,
-                records=records,
-                result=driver_result,
-            ),
-        )
-
-    def _deployment_record_id(self, provider_operation_key: str) -> str:
-        operation_digest = hashlib.sha256(provider_operation_key.encode("utf-8")).hexdigest()[:24]
-        return (
-            f"deployment-provider-operation-{operation_digest}-"
-            f"{self._lane.context}-{self._lane.instance}"
-        )
-
-    def apply(
-        self, provider_operation_key: str, lease: ProviderOperationLease
-    ) -> ProviderMutationOutcome:
-        try:
-            records, result = execute_generic_web_deploy_result(
-                control_plane_root=self._control_plane_root,
-                record_store=self._record_store,
-                request=self._deploy_request,
-                profile=self._profile,
-                lane=self._lane,
-                provider_operation_title=provider_operation_title(provider_operation_key),
-                deployment_record_id=self._deployment_record_id(provider_operation_key),
-                deploy_provider=self._deploy_provider,
-                resolved_deploy_target=self._resolve_deploy_target(),
-                provider_effect_checkpoint=lease.checkpoint_effect,
-            )
-        except (FileNotFoundError, ValueError) as error:
-            raise ProviderMutationRejectedError(error)
-        except click.ClickException as error:
-            raise ProviderMutationUnknownError(str(error)) from error
-        provider_effect_attempted = result.pop("provider_effect_attempted", False) is True
-        if str(result.get("deploy_status", "")).strip() == "fail" and provider_effect_attempted:
-            raise ProviderMutationUnknownError(
-                str(result.get("error_message", "")).strip()
-                or "Generic web provider outcome requires reconciliation."
-            )
-        return ProviderMutationOutcome(
-            response_status_code=202,
-            response_payload=_provider_operation_response_payload(
-                trace_id=self._trace_id,
-                records=records,
-                result=result,
-            ),
-            durable=should_store_generic_web_deploy_idempotency(result),
-            provider_effect_performed=provider_effect_attempted,
-        )
-
-
-class _GenericWebProdPromotionProviderMutationAdapter:
-    def __init__(
-        self,
-        *,
-        control_plane_root: FilePath,
-        record_store: object,
-        promotion_request: GenericWebProdPromotionEnvelope,
-        profile: LaunchplaneProductProfileRecord,
-        lane: ProductLaneProfile,
-        trace_id: str,
-        validate_before_effect: Callable[[GenericWebResolvedDeployTarget], None],
-    ) -> None:
-        self._control_plane_root = control_plane_root
-        self._record_store = record_store
-        self._promotion_request = promotion_request
-        self._profile = profile
-        self._lane = lane
-        self._trace_id = trace_id
-        self._validate_before_effect = validate_before_effect
-        self._deploy_provider: GenericWebDeployProvider = default_generic_web_deploy_provider()
-        self._resolved_deploy_target: GenericWebResolvedDeployTarget | None = None
-
-    def resolve_deploy_target(self) -> GenericWebResolvedDeployTarget:
-        if self._resolved_deploy_target is None:
-            promotion = self._promotion_request.promotion
-            self._resolved_deploy_target = self._deploy_provider.resolve_deploy_target(
-                control_plane_root=self._control_plane_root,
-                request_artifact_id=promotion.artifact_id,
-                request_source_git_ref=promotion.source_git_ref,
-                request_timeout_seconds=promotion.timeout_seconds,
-                request_no_cache=promotion.no_cache,
-                record_store=self._record_store,
-                profile=self._profile,
-                lane=self._lane,
-                normalized_artifact_id=normalize_generic_web_artifact_id(
-                    profile=self._profile,
-                    artifact_id=promotion.artifact_id,
-                ),
-                fallback_target_name=f"{self._profile.product}-{self._lane.instance}",
-            )
-        return self._resolved_deploy_target
-
-    def reconciliation_key(self) -> str:
-        return build_generic_web_provider_reconciliation_key(self.resolve_deploy_target())
-
-    def target_key(self) -> str:
-        return build_generic_web_provider_target_key(self.resolve_deploy_target())
-
-    def observe(
-        self,
-        provider_operation_key: str,
-        provider_effect_phase: str,
-        reconciliation_key: str,
-    ) -> ProviderObservation:
-        del provider_operation_key, provider_effect_phase, reconciliation_key
-        return ProviderObservation(outcome="unknown")
-
-    def _deployment_record_id(self, provider_operation_key: str) -> str:
-        operation_digest = hashlib.sha256(provider_operation_key.encode("utf-8")).hexdigest()[:24]
-        return (
-            f"deployment-provider-operation-{operation_digest}-"
-            f"{self._lane.context}-{self._lane.instance}"
-        )
-
-    def apply(
-        self, provider_operation_key: str, lease: ProviderOperationLease
-    ) -> ProviderMutationOutcome:
-        provider_effect_attempted = False
-
-        def checkpoint_provider_effect(phase: str) -> None:
-            nonlocal provider_effect_attempted
-            lease.checkpoint_effect(phase)
-            provider_effect_attempted = True
-
-        try:
-            self._validate_before_effect(self.resolve_deploy_target())
-            records, result = execute_generic_web_prod_promotion_result(
-                control_plane_root=self._control_plane_root,
-                record_store=self._record_store,
-                request=self._promotion_request,
-                deploy_provider=self._deploy_provider,
-                resolved_deploy_target=self.resolve_deploy_target(),
-                provider_operation_title=provider_operation_title(provider_operation_key),
-                deployment_record_id=self._deployment_record_id(provider_operation_key),
-                provider_effect_checkpoint=checkpoint_provider_effect,
-            )
-        except StarletteHTTPException as error:
-            raise ProviderMutationRejectedError(error) from error
-        except (FileNotFoundError, ValueError, click.ClickException) as error:
-            if provider_effect_attempted:
-                raise ProviderMutationUnknownError(str(error)) from error
-            raise ProviderMutationRejectedError(error) from error
-        return ProviderMutationOutcome(
-            response_status_code=202,
-            response_payload=_provider_operation_response_payload(
-                trace_id=self._trace_id,
-                records=records.model_dump(mode="json"),
-                result=result.model_dump(mode="json"),
-            ),
-            durable=should_store_generic_web_promotion_idempotency(result),
             provider_effect_performed=provider_effect_attempted,
         )
 
@@ -2715,32 +2343,6 @@ def resolve_preview_pr_feedback_context(*, record_store: object, product: str, c
     if requested_context and requested_context != profile_context:
         raise ValueError("Preview PR feedback context does not match product profile.")
     return profile_context
-
-
-def read_generic_web_preview_profile(
-    *, record_store: object, product: str
-) -> LaunchplaneProductProfileRecord:
-    try:
-        profile_store = require_product_profile_read_store(record_store)
-        profile = profile_store.read_product_profile_record(product.strip())
-    except TypeError as error:
-        raise TypeError(
-            "Launchplane record store does not support generic web preview desired-state "
-            "profile reads: read_product_profile_record"
-        ) from error
-    except FileNotFoundError as error:
-        raise FileNotFoundError(
-            "Generic web preview desired-state requires an existing product profile."
-        ) from error
-    if not product_profile_uses_generic_web_base(profile):
-        raise ValueError(
-            "Product profile is not compatible with the generic-web preview desired-state route."
-        )
-    if not profile.preview.enabled:
-        raise ValueError("Product profile does not have generic-web previews enabled.")
-    if not profile.preview.context.strip():
-        raise ValueError("Product profile does not define a preview context.")
-    return profile
 
 
 def require_ingress_edge_endpoint_read_store(record_store: object) -> _IngressEdgeEndpointReadStore:
@@ -6449,256 +6051,6 @@ def create_launchplane_fastapi_app(
             records={key: str(value) for key, value in records.items()},
             result=driver_result,
         )
-
-    async def write_generic_web_rollback_plan(
-        request: Request,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            raw_payload = await request.json()
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request payload failed validation.",
-            ) from error
-        if not isinstance(raw_payload, dict):
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request payload failed validation.",
-            )
-        try:
-            rollback_request = GenericWebRollbackPlanEnvelope.model_validate(raw_payload)
-        except ValidationError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request payload failed validation.",
-            ) from error
-        try:
-            _profile, lane = resolve_generic_web_rollback_lane(
-                record_store=record_store,
-                product=rollback_request.product,
-                route_path=_GENERIC_WEB_ROLLBACK_PLAN_ROUTE,
-                instance=rollback_request.rollback_plan.instance,
-            )
-        except GenericWebRollbackRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_ROLLBACK_PLAN_ROUTE,
-            )
-        except GenericWebRollbackProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(write_generic_web_rollback_plan),
-            product=rollback_request.product,
-            context=lane.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot plan the generic web prod rollback"
-                    " for the requested product/context."
-                ),
-            )
-        (
-            normalized_idempotency_key,
-            payload_fingerprint,
-            replay_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_ROLLBACK_PLAN_ROUTE,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=bool(idempotency_key.strip()),
-        )
-        if replay_response is not None:
-            return replay_response
-        try:
-            records, driver_result = execute_generic_web_rollback_plan_result(
-                record_store=record_store,
-                request=rollback_request,
-            )
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=f"No Launchplane route for {_GENERIC_WEB_ROLLBACK_PLAN_ROUTE}.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        response = accepted_evidence_response(
-            trace_id=trace_id,
-            records={key: str(value) for key, value in records.items()},
-            result=driver_result,
-        )
-        if should_store_generic_web_rollback_idempotency(driver_result):
-            store_apply_idempotency(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_ROLLBACK_PLAN_ROUTE,
-                idempotency_key=normalized_idempotency_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        return response
-
-    async def write_generic_web_rollback(
-        request: Request,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            raw_payload = await request.json()
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request payload failed validation.",
-            ) from error
-        if not isinstance(raw_payload, dict):
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request payload failed validation.",
-            )
-        try:
-            rollback_request = GenericWebRollbackEnvelope.model_validate(raw_payload)
-        except ValidationError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request payload failed validation.",
-            ) from error
-        try:
-            profile, lane = resolve_generic_web_rollback_lane(
-                record_store=record_store,
-                product=rollback_request.product,
-                route_path=_GENERIC_WEB_ROLLBACK_ROUTE,
-                instance=rollback_request.rollback.instance,
-            )
-        except GenericWebRollbackRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_ROLLBACK_ROUTE,
-            )
-        except GenericWebRollbackProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(write_generic_web_rollback),
-            product=rollback_request.product,
-            context=lane.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot execute the generic web prod rollback"
-                    " for the requested product/context."
-                ),
-            )
-        (
-            normalized_idempotency_key,
-            payload_fingerprint,
-            replay_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_ROLLBACK_ROUTE,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=bool(idempotency_key.strip()),
-        )
-        if replay_response is not None:
-            return replay_response
-        try:
-            records, driver_result = execute_generic_web_rollback_result(
-                control_plane_root=resolved_control_plane_root,
-                record_store=record_store,
-                request=rollback_request,
-                profile=profile,
-            )
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=f"No Launchplane route for {_GENERIC_WEB_ROLLBACK_ROUTE}.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        response = accepted_evidence_response(
-            trace_id=trace_id,
-            records={key: str(value) for key, value in records.items()},
-            result=driver_result,
-        )
-        if should_store_generic_web_rollback_idempotency(driver_result):
-            store_apply_idempotency(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_ROLLBACK_ROUTE,
-                idempotency_key=normalized_idempotency_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        return response
 
     async def write_odoo_prod_rollback(
         request: Request,
@@ -10444,107 +9796,6 @@ def create_launchplane_fastapi_app(
                 message="Current product promotion evidence is invalid.",
             ) from error
 
-    def require_product_promotion_intent(
-        *,
-        record_store: object,
-        profile: LaunchplaneProductProfileRecord,
-        lane: ProductLaneProfile,
-        promotion_request: GenericWebProdPromotionEnvelope,
-        idempotency_key: str,
-        trace_id: str,
-    ) -> None:
-        intent_id = promotion_request.promotion.promotion_intent_id.strip()
-        if not isinstance(record_store, PostgresRecordStore) or not intent_id:
-            raise _launchplane_http_error(
-                status_code=409,
-                trace_id=trace_id,
-                code="promotion_intent_required",
-                message="Live generic-web promotion requires a product-owned promotion intent.",
-            )
-        if idempotency_key.strip() != intent_id:
-            raise _launchplane_http_error(
-                status_code=409,
-                trace_id=trace_id,
-                code="promotion_intent_idempotency_mismatch",
-                message="Live promotion must use its promotion intent as Idempotency-Key.",
-            )
-        try:
-            intent_record = record_store.read_outbox_delivery_record(intent_id)
-            current_profile, current_lane, status = build_product_promotion_status(
-                record_store=record_store,
-                product=profile.product,
-                destination_environment=lane.instance,
-                action_allowed=lambda _action, _product, _context: True,
-                workflow_credentials_ready=lambda _context: True,
-            )
-        except (AttributeError, FileNotFoundError, ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=409,
-                trace_id=trace_id,
-                code="promotion_intent_invalid",
-                message="Product-owned promotion intent is unavailable or stale.",
-            ) from error
-        if (
-            current_profile != profile
-            or current_lane != lane
-            or not status.workflow_live.enabled
-            or not product_promotion_intent_matches(
-                record=intent_record,
-                profile=current_profile,
-                status=status,
-                request=promotion_request.promotion,
-            )
-        ):
-            raise _launchplane_http_error(
-                status_code=409,
-                trace_id=trace_id,
-                code="promotion_intent_invalid",
-                message="Product-owned promotion intent does not match current evidence.",
-            )
-
-    def require_product_promotion_target_snapshot(
-        *,
-        record_store: object,
-        lane: ProductLaneProfile,
-        resolved_deploy_target: GenericWebResolvedDeployTarget,
-        trace_id: str,
-    ) -> None:
-        if not isinstance(record_store, PostgresRecordStore):
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message="Live generic-web promotion requires Launchplane database storage.",
-            )
-        try:
-            current_target = record_store.read_provider_target_record(
-                context_name=lane.context,
-                instance_name=lane.instance,
-            )
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=409,
-                trace_id=trace_id,
-                code="promotion_target_changed",
-                message="The reviewed production provider target is no longer available.",
-            ) from error
-        expected_target = resolved_deploy_target.deployed_target
-        resolved_target = resolved_deploy_target.resolved_target
-        if (
-            expected_target is None
-            or current_target.to_deployed_target_reference() != expected_target
-            or resolved_target.target_id != expected_target.target_id
-            or resolved_target.target_name != expected_target.display_name
-            or resolved_target.target_type
-            != (expected_target.provider_target_type or expected_target.target_category)
-        ):
-            raise _launchplane_http_error(
-                status_code=409,
-                trace_id=trace_id,
-                code="promotion_target_changed",
-                message="The production provider target changed before promotion execution.",
-            )
-
     async def dry_run_product_promotion(
         product: str,
         environment: str,
@@ -13259,896 +12510,6 @@ def create_launchplane_fastapi_app(
             )
         return response
 
-    async def apply_generic_web_preview_desired_state(
-        request: Request,
-        desired_state_request: GenericWebPreviewDesiredStateEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> AcceptedEvidenceResponse:
-        trace_id = next_trace_id()
-        try:
-            profile = read_generic_web_preview_profile(
-                record_store=record_store,
-                product=desired_state_request.product,
-            )
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="driver_route_dependency_not_found",
-                message=(
-                    "Driver route is registered, but required product or runtime records "
-                    "were not found."
-                ),
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                apply_generic_web_preview_desired_state
-            ),
-            product=profile.product,
-            context=profile.preview.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot discover generic web preview desired state "
-                    "for the requested product/context."
-                ),
-            )
-        (
-            normalized_key,
-            payload_fingerprint,
-            replayed_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_PREVIEW_DESIRED_STATE_ROUTE.route_path,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=True,
-        )
-        if replayed_response is not None:
-            return replayed_response
-        try:
-            desired_state_store = require_preview_desired_state_write_store(record_store)
-        except TypeError as error:
-            raise _launchplane_http_error(
-                status_code=503,
-                trace_id=trace_id,
-                code="database_storage_required",
-                message=str(error),
-            ) from error
-        driver_result = discover_generic_web_preview_desired_state(
-            control_plane_root=resolved_control_plane_root,
-            record_store=cast(GenericWebPreviewProfileStore, record_store),
-            request=desired_state_request.desired_state,
-            discovered_at=utc_now_timestamp(),
-            profile=profile,
-        )
-        desired_state_store.write_preview_desired_state_record(driver_result)
-        response = accepted_evidence_response(
-            trace_id=trace_id,
-            records={"preview_desired_state_id": driver_result.desired_state_id},
-            result=driver_result.model_dump(mode="json"),
-        )
-        if driver_result.status != "fail":
-            store_apply_idempotency(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_PREVIEW_DESIRED_STATE_ROUTE.route_path,
-                idempotency_key=normalized_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        return response
-
-    async def apply_generic_web_preview_inventory(
-        inventory_request: GenericWebPreviewInventoryEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            profile = resolve_generic_web_preview_profile(
-                record_store=record_store,
-                product=inventory_request.product,
-            )
-        except GenericWebPreviewRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_PREVIEW_INVENTORY_ROUTE,
-            )
-        except GenericWebPreviewProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                apply_generic_web_preview_inventory
-            ),
-            product=profile.product,
-            context=profile.preview.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot read generic web preview inventory"
-                    " for the requested product/context."
-                ),
-            )
-        try:
-            records, result = apply_generic_web_preview_inventory_result(
-                control_plane_root=resolved_control_plane_root,
-                record_store=record_store,
-                request=inventory_request,
-                profile=profile,
-            )
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        return accepted_evidence_response(
-            trace_id=trace_id,
-            records=records,
-            result=result,
-        )
-
-    async def apply_generic_web_preview_readiness(
-        readiness_request: GenericWebPreviewReadinessEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            profile = resolve_generic_web_preview_profile(
-                record_store=record_store,
-                product=readiness_request.product,
-            )
-        except GenericWebPreviewRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_PREVIEW_READINESS_ROUTE,
-            )
-        except GenericWebPreviewProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                apply_generic_web_preview_readiness
-            ),
-            product=profile.product,
-            context=profile.preview.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot evaluate generic web preview readiness"
-                    " for the requested product/context."
-                ),
-            )
-        try:
-            records, result = apply_generic_web_preview_readiness_result(
-                control_plane_root=resolved_control_plane_root,
-                record_store=record_store,
-                request=readiness_request,
-                profile=profile,
-            )
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        return accepted_evidence_response(
-            trace_id=trace_id,
-            records=records,
-            result=result,
-        )
-
-    async def apply_generic_web_preview_refresh(
-        request: Request,
-        refresh_request: GenericWebPreviewRefreshEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            profile = resolve_generic_web_preview_profile(
-                record_store=record_store,
-                product=refresh_request.product,
-            )
-        except GenericWebPreviewRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_PREVIEW_REFRESH_ROUTE,
-            )
-        except GenericWebPreviewProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                apply_generic_web_preview_refresh
-            ),
-            product=profile.product,
-            context=profile.preview.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot refresh generic web preview state"
-                    " for the requested product/context."
-                ),
-            )
-        (
-            normalized_key,
-            payload_fingerprint,
-            replayed_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_PREVIEW_REFRESH_ROUTE,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=bool(idempotency_key.strip()),
-        )
-        if replayed_response is not None:
-            return replayed_response
-        try:
-            records, result = apply_generic_web_preview_refresh_result(
-                control_plane_root=resolved_control_plane_root,
-                record_store=record_store,
-                request=refresh_request,
-                profile=profile,
-            )
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        response = accepted_evidence_response(
-            trace_id=trace_id,
-            records=records,
-            result=result,
-        )
-        if should_store_generic_web_preview_idempotency(result):
-            store_apply_idempotency(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_PREVIEW_REFRESH_ROUTE,
-                idempotency_key=normalized_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        return response
-
-    async def apply_generic_web_preview_destroy(
-        request: Request,
-        destroy_request: GenericWebPreviewDestroyEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            profile = resolve_generic_web_preview_profile(
-                record_store=record_store,
-                product=destroy_request.product,
-            )
-        except GenericWebPreviewRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_PREVIEW_DESTROY_ROUTE,
-            )
-        except GenericWebPreviewProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                apply_generic_web_preview_destroy
-            ),
-            product=profile.product,
-            context=profile.preview.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot destroy generic web preview state"
-                    " for the requested product/context."
-                ),
-            )
-        (
-            normalized_key,
-            payload_fingerprint,
-            replayed_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_PREVIEW_DESTROY_ROUTE,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=bool(idempotency_key.strip()),
-        )
-        if replayed_response is not None:
-            return replayed_response
-        try:
-            records, result = apply_generic_web_preview_destroy_result(
-                control_plane_root=resolved_control_plane_root,
-                record_store=record_store,
-                request=destroy_request,
-                profile=profile,
-            )
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        response = accepted_evidence_response(
-            trace_id=trace_id,
-            records=records,
-            result=result,
-        )
-        if should_store_generic_web_preview_idempotency(result):
-            store_apply_idempotency(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_PREVIEW_DESTROY_ROUTE,
-                idempotency_key=normalized_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        return response
-
-    async def apply_generic_web_deploy(
-        request: Request,
-        deploy_request: GenericWebDeployEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            profile, lane = resolve_generic_web_deploy_lane(
-                record_store=record_store,
-                product=deploy_request.product,
-                instance=deploy_request.deploy.instance,
-            )
-        except GenericWebDeployRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_DEPLOY_ROUTE,
-            )
-        except GenericWebDeployProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(apply_generic_web_deploy),
-            product=profile.product,
-            context=lane.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot execute the generic web deploy driver"
-                    " for the requested product/context."
-                ),
-            )
-        normalized_key = idempotency_key.strip()
-        if not normalized_key:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="idempotency_key_required",
-                message="Generic web deploy requests require an Idempotency-Key header.",
-            )
-        raw_payload = await request.json()
-        payload_fingerprint = idempotency_request_fingerprint(
-            route_path=_GENERIC_WEB_DEPLOY_ROUTE,
-            payload=cast(dict[str, object], raw_payload),
-        )
-        adapter = _GenericWebDeployProviderMutationAdapter(
-            control_plane_root=resolved_control_plane_root,
-            record_store=record_store,
-            deploy_request=deploy_request,
-            profile=profile,
-            lane=lane,
-            trace_id=trace_id,
-        )
-        try:
-            return await run_provider_mutation(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_DEPLOY_ROUTE,
-                idempotency_key=normalized_key,
-                request_fingerprint=payload_fingerprint,
-                trace_id=trace_id,
-                adapter=adapter,
-                in_progress_message=(
-                    "A matching generic web deploy is already running. "
-                    "Retry with the same Idempotency-Key."
-                ),
-                reconcile_message=("The generic web deploy requires reconciliation before retry."),
-            )
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=f"No Launchplane route for {_GENERIC_WEB_DEPLOY_ROUTE}.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-
-    async def apply_generic_web_prod_promotion(
-        request: Request,
-        promotion_request: GenericWebProdPromotionEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_browser_mutation_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> GenericWebProdPromotionResponse | JSONResponse:
-        trace_id = next_trace_id()
-        if isinstance(identity, TerminalAgentIdentity):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message="Terminal agent credentials can only read redacted Launchplane context.",
-            )
-        try:
-            profile, lane = resolve_generic_web_promotion_destination_lane(
-                record_store=record_store,
-                product=promotion_request.product,
-                instance=promotion_request.promotion.to_instance,
-            )
-            validate_generic_web_prod_promotion_lanes(
-                record_store=record_store,
-                request=promotion_request,
-            )
-        except GenericWebPromotionRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_PROD_PROMOTION_ROUTE,
-            )
-        except GenericWebPromotionProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if (
-            isinstance(identity, GitHubHumanIdentity | LocalOperatorIdentity | LocalAdminIdentity)
-            and not promotion_request.promotion.dry_run
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message="Launchplane UI can only dry-run generic-web prod promotions.",
-            )
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                apply_generic_web_prod_promotion
-            ),
-            product=profile.product,
-            context=lane.context.strip(),
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot execute the generic web prod promotion driver"
-                    " for the requested product/context."
-                ),
-            )
-        live_promotion = not promotion_request.promotion.dry_run
-        promotion_intent_id = promotion_request.promotion.promotion_intent_id.strip()
-        if live_promotion:
-            if not promotion_intent_id and not (
-                resolved_authz_policy_runtime.policy.allows(
-                    identity=identity,
-                    action="generic_web_prod_promotion.execute_unreviewed",
-                    product=profile.product,
-                    context=lane.context.strip(),
-                )
-            ):
-                raise _launchplane_http_error(
-                    status_code=403,
-                    trace_id=trace_id,
-                    code="promotion_intent_required",
-                    message=(
-                        "Live generic-web promotion requires a product-owned intent or an "
-                        "explicit unreviewed-automation grant."
-                    ),
-                )
-            if not idempotency_key.strip():
-                raise _launchplane_http_error(
-                    status_code=400,
-                    trace_id=trace_id,
-                    code="idempotency_key_required",
-                    message="Live generic-web promotion requires an Idempotency-Key header.",
-                )
-        reservation_scope = (
-            f"promotion-intent:{promotion_intent_id}"
-            if live_promotion and promotion_intent_id
-            else ""
-        )
-        (
-            normalized_key,
-            payload_fingerprint,
-            replayed_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_PROD_PROMOTION_ROUTE,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=bool(idempotency_key.strip()),
-            idempotency_scope_override=reservation_scope,
-        )
-        if replayed_response is not None:
-            return GenericWebProdPromotionResponse.model_validate(
-                replayed_response.model_dump(mode="json")
-            )
-        if live_promotion:
-
-            def validate_live_promotion(
-                resolved_deploy_target: GenericWebResolvedDeployTarget,
-            ) -> None:
-                require_product_promotion_target_snapshot(
-                    record_store=record_store,
-                    lane=lane,
-                    resolved_deploy_target=resolved_deploy_target,
-                    trace_id=trace_id,
-                )
-                if promotion_intent_id:
-                    require_product_promotion_intent(
-                        record_store=record_store,
-                        profile=profile,
-                        lane=lane,
-                        promotion_request=promotion_request,
-                        idempotency_key=idempotency_key,
-                        trace_id=trace_id,
-                    )
-
-            adapter = _GenericWebProdPromotionProviderMutationAdapter(
-                control_plane_root=resolved_control_plane_root,
-                record_store=record_store,
-                promotion_request=promotion_request,
-                profile=profile,
-                lane=lane,
-                trace_id=trace_id,
-                validate_before_effect=validate_live_promotion,
-            )
-            try:
-                durable_response = await run_provider_mutation(
-                    record_store=record_store,
-                    identity=identity,
-                    route_path=_GENERIC_WEB_PROD_PROMOTION_ROUTE,
-                    idempotency_key=normalized_key,
-                    request_fingerprint=payload_fingerprint,
-                    trace_id=trace_id,
-                    adapter=adapter,
-                    in_progress_message=(
-                        "A matching generic web prod promotion is already running. "
-                        "Retry with the same Idempotency-Key."
-                    ),
-                    reconcile_message=(
-                        "The generic web prod promotion requires reconciliation before retry."
-                    ),
-                    reservation_scope=reservation_scope,
-                )
-            except FileNotFoundError as error:
-                raise _launchplane_http_error(
-                    status_code=404,
-                    trace_id=trace_id,
-                    code="not_found",
-                    message=f"No Launchplane route for {_GENERIC_WEB_PROD_PROMOTION_ROUTE}.",
-                ) from error
-            except (ValueError, click.ClickException) as error:
-                raise _launchplane_http_error(
-                    status_code=400,
-                    trace_id=trace_id,
-                    code="invalid_request",
-                    message="Request could not be completed.",
-                ) from error
-            return GenericWebProdPromotionResponse.model_validate(
-                durable_response.model_dump(mode="json")
-            )
-        try:
-            records, result = execute_generic_web_prod_promotion_result(
-                control_plane_root=resolved_control_plane_root,
-                record_store=record_store,
-                request=promotion_request,
-            )
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=f"No Launchplane route for {_GENERIC_WEB_PROD_PROMOTION_ROUTE}.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        response = GenericWebProdPromotionResponse(
-            trace_id=trace_id,
-            records=records,
-            result=GenericWebProdPromotionResponseResult.model_validate(
-                result.model_dump(mode="json")
-            ),
-        )
-        if should_store_generic_web_promotion_idempotency(result):
-            store_apply_idempotency(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_PROD_PROMOTION_ROUTE,
-                idempotency_key=normalized_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        return response
-
-    async def dispatch_generic_web_prod_promotion_workflow(
-        request: Request,
-        workflow_request: GenericWebPromotionWorkflowEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_browser_mutation_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> GenericWebPromotionWorkflowResponse | JSONResponse:
-        trace_id = next_trace_id()
-        if isinstance(identity, TerminalAgentIdentity):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message="Terminal agent credentials can only read redacted Launchplane context.",
-            )
-        if isinstance(identity, GitHubHumanIdentity | LocalOperatorIdentity | LocalAdminIdentity):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Operator workflow dispatches must use the product-owned promotion route."
-                ),
-            )
-        try:
-            profile, lane = resolve_generic_web_promotion_workflow_lane(
-                record_store=record_store,
-                product=workflow_request.product,
-                context=workflow_request.workflow.context,
-            )
-        except GenericWebPromotionRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_PROD_PROMOTION_WORKFLOW_ROUTE,
-            )
-        except GenericWebPromotionProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                dispatch_generic_web_prod_promotion_workflow
-            ),
-            product=profile.product,
-            context=lane.context.strip(),
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot dispatch the generic web prod promotion workflow"
-                    " for the requested product/context."
-                ),
-            )
-        (
-            normalized_key,
-            payload_fingerprint,
-            replayed_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_PROD_PROMOTION_WORKFLOW_ROUTE,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=bool(idempotency_key.strip()),
-        )
-        if replayed_response is not None:
-            return GenericWebPromotionWorkflowResponse.model_validate(
-                replayed_response.model_dump(mode="json")
-            )
-        try:
-            records, result, outbox_delivery = dispatch_generic_web_promotion_workflow_result(
-                request=workflow_request,
-                profile=profile,
-                delivery_key=(
-                    f"{idempotency_scope(identity)}:{normalized_key}"
-                    if normalized_key
-                    else trace_id
-                ),
-            )
-        except FileNotFoundError as error:
-            raise _launchplane_http_error(
-                status_code=404,
-                trace_id=trace_id,
-                code="not_found",
-                message=(f"No Launchplane route for {_GENERIC_WEB_PROD_PROMOTION_WORKFLOW_ROUTE}."),
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        response = GenericWebPromotionWorkflowResponse(
-            trace_id=trace_id,
-            records=records,
-            result=GenericWebPromotionWorkflowResponseResult.model_validate(
-                result.model_dump(mode="json")
-            ),
-        )
-        if not isinstance(record_store, PostgresRecordStore):
-            raise _launchplane_http_error(
-                status_code=409,
-                trace_id=trace_id,
-                code="outbox_requires_postgres",
-                message="Workflow dispatch requires PostgreSQL transactional outbox storage.",
-            )
-        idempotency_record = None
-        if normalized_key and should_store_generic_web_promotion_idempotency(result):
-            idempotency_record = build_apply_idempotency_record(
-                identity=identity,
-                route_path=_GENERIC_WEB_PROD_PROMOTION_WORKFLOW_ROUTE,
-                idempotency_key=normalized_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        record_store.enqueue_outbox_delivery_with_idempotency(
-            OutboxWithIdempotencyRequest(
-                delivery=outbox_delivery,
-                idempotency_record=idempotency_record,
-            )
-        )
-        return response
-
     def resolve_verireel_route_authorization(
         *,
         record_store: object,
@@ -15304,197 +13665,6 @@ def create_launchplane_fastapi_app(
             )
         return response
 
-    async def apply_generic_web_stable_verification(
-        request: Request,
-        verification_request: GenericWebStableVerificationEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            profile, lane = resolve_generic_web_stable_verification_lane(
-                record_store=record_store,
-                product=verification_request.product,
-                context=verification_request.verification.context,
-                instance=verification_request.verification.instance,
-            )
-        except GenericWebVerificationRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_STABLE_VERIFICATION_ROUTE,
-            )
-        except GenericWebVerificationProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except ValueError as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                apply_generic_web_stable_verification
-            ),
-            product=profile.product,
-            context=lane.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot write generic web stable verification"
-                    " for the requested product/context."
-                ),
-            )
-        (
-            normalized_key,
-            payload_fingerprint,
-            replayed_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_STABLE_VERIFICATION_ROUTE,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=bool(idempotency_key.strip()),
-        )
-        if replayed_response is not None:
-            return replayed_response
-        try:
-            result = apply_generic_web_stable_verification_result(
-                record_store=record_store,
-                request=verification_request,
-            )
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        response = accepted_evidence_response(
-            trace_id=trace_id,
-            records=generic_web_verification_response_records(result),
-            result=result,
-        )
-        if should_store_generic_web_verification_idempotency(result):
-            store_apply_idempotency(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_STABLE_VERIFICATION_ROUTE,
-                idempotency_key=normalized_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        return response
-
-    async def apply_generic_web_preview_verification(
-        request: Request,
-        verification_request: GenericWebPreviewVerificationEnvelope,
-        identity: Annotated[LaunchplaneIdentity, Depends(read_write_identity)],
-        record_store: Annotated[object, Depends(get_record_store)],
-        idempotency_key: Annotated[str, Header(alias="Idempotency-Key")] = "",
-    ) -> AcceptedEvidenceResponse | JSONResponse:
-        trace_id = next_trace_id()
-        try:
-            profile = resolve_generic_web_preview_verification_profile(
-                record_store=record_store,
-                product=verification_request.product,
-                context=verification_request.verification.context,
-            )
-        except GenericWebVerificationRouteDependencyError:
-            return driver_route_dependency_not_found_response(
-                trace_id=trace_id,
-                route_path=_GENERIC_WEB_PREVIEW_VERIFICATION_ROUTE,
-            )
-        except GenericWebVerificationProductMismatchError as error:
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="product_driver_mismatch",
-                message="Product is not configured for the requested driver route.",
-            ) from error
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        if not resolved_authz_policy_runtime.policy.allows(
-            identity=identity,
-            action=native_routes._native_driver_route_authz_action(
-                apply_generic_web_preview_verification
-            ),
-            product=profile.product,
-            context=profile.preview.context,
-        ):
-            raise _launchplane_http_error(
-                status_code=403,
-                trace_id=trace_id,
-                code="authorization_denied",
-                message=(
-                    "Workflow cannot write generic web preview verification"
-                    " for the requested product/context."
-                ),
-            )
-        (
-            normalized_key,
-            payload_fingerprint,
-            replayed_response,
-        ) = await replay_apply_idempotency(
-            request=request,
-            record_store=record_store,
-            identity=identity,
-            route_path=_GENERIC_WEB_PREVIEW_VERIFICATION_ROUTE,
-            idempotency_key=idempotency_key,
-            trace_id=trace_id,
-            check_replay=bool(idempotency_key.strip()),
-        )
-        if replayed_response is not None:
-            return replayed_response
-        verification_request.verification.context = profile.preview.context
-        try:
-            result = apply_generic_web_preview_verification_result(
-                control_plane_root=resolved_control_plane_root,
-                record_store=record_store,
-                request=verification_request,
-            )
-        except (ValueError, click.ClickException) as error:
-            raise _launchplane_http_error(
-                status_code=400,
-                trace_id=trace_id,
-                code="invalid_request",
-                message="Request could not be completed.",
-            ) from error
-        response = accepted_evidence_response(
-            trace_id=trace_id,
-            records=generic_web_verification_response_records(result),
-            result=result,
-        )
-        if should_store_generic_web_verification_idempotency(result):
-            store_apply_idempotency(
-                record_store=record_store,
-                identity=identity,
-                route_path=_GENERIC_WEB_PREVIEW_VERIFICATION_ROUTE,
-                idempotency_key=normalized_key,
-                request_fingerprint_value=payload_fingerprint,
-                trace_id=trace_id,
-                response=response,
-            )
-        return response
-
     async def apply_ingress_canary_route(
         request: Request,
         canary_request: IngressCanaryRouteApplyEnvelope,
@@ -16374,273 +14544,31 @@ def create_launchplane_fastapi_app(
         },
     )
 
-    app.add_api_route(
-        _GENERIC_WEB_PREVIEW_DESIRED_STATE_ROUTE.route_path,
-        apply_generic_web_preview_desired_state,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        operation_id="apply_generic_web_preview_desired_state",
-        summary="Discover generic web preview desired state",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
+    generic_web_write_route_dependencies = GenericWebWriteRouteDependencies(
+        read_write_identity=read_write_identity,
+        read_browser_mutation_identity=read_browser_mutation_identity,
+        get_record_store=get_record_store,
+        next_trace_id=next_trace_id,
+        authorization_allows=resolved_authz_policy_runtime.policy.allows,
+        http_error=_launchplane_http_error,
+        error_response_model=LaunchplaneErrorResponse,
+        control_plane_root=resolved_control_plane_root,
+        driver_dependency_response=driver_route_dependency_not_found_response,
+        replay_apply_idempotency=replay_apply_idempotency,
+        store_apply_idempotency=store_apply_idempotency,
+        build_apply_idempotency_record=build_apply_idempotency_record,
+        run_provider_mutation=run_provider_mutation,
+        idempotency_request_fingerprint=idempotency_request_fingerprint,
+        require_preview_desired_state_write_store=require_preview_desired_state_write_store,
+        openapi_model_schema=_openapi_model_schema,
     )
-
-    app.add_api_route(
-        _GENERIC_WEB_PREVIEW_INVENTORY_ROUTE,
-        apply_generic_web_preview_inventory,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebPreviewInventoryEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="apply_generic_web_preview_inventory",
-        summary="Read generic web preview inventory",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
+    generic_web_write_route_handlers = build_generic_web_write_route_handlers(
+        dependencies=generic_web_write_route_dependencies,
     )
-
-    app.add_api_route(
-        _GENERIC_WEB_PREVIEW_READINESS_ROUTE,
-        apply_generic_web_preview_readiness,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebPreviewReadinessEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="apply_generic_web_preview_readiness",
-        summary="Evaluate generic web preview readiness",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        _GENERIC_WEB_PREVIEW_REFRESH_ROUTE,
-        apply_generic_web_preview_refresh,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebPreviewRefreshEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="apply_generic_web_preview_refresh",
-        summary="Refresh generic web preview",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        _GENERIC_WEB_PREVIEW_DESTROY_ROUTE,
-        apply_generic_web_preview_destroy,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebPreviewDestroyEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="apply_generic_web_preview_destroy",
-        summary="Destroy generic web preview",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        _GENERIC_WEB_DEPLOY_ROUTE,
-        apply_generic_web_deploy,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {"schema": _openapi_model_schema(GenericWebDeployEnvelope)}
-                },
-            }
-        },
-        operation_id="apply_generic_web_deploy",
-        summary="Execute generic web deploy",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        _GENERIC_WEB_PROD_PROMOTION_ROUTE,
-        apply_generic_web_prod_promotion,
-        methods=["POST"],
-        status_code=202,
-        response_model=GenericWebProdPromotionResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebProdPromotionEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="apply_generic_web_prod_promotion",
-        summary="Execute generic web prod promotion",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        _GENERIC_WEB_PROD_PROMOTION_WORKFLOW_ROUTE,
-        dispatch_generic_web_prod_promotion_workflow,
-        methods=["POST"],
-        status_code=202,
-        response_model=GenericWebPromotionWorkflowResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebPromotionWorkflowEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="dispatch_generic_web_prod_promotion_workflow",
-        summary="Dispatch generic web prod promotion workflow",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        _GENERIC_WEB_STABLE_VERIFICATION_ROUTE,
-        apply_generic_web_stable_verification,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebStableVerificationEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="apply_generic_web_stable_verification",
-        summary="Write generic web stable verification",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        _GENERIC_WEB_PREVIEW_VERIFICATION_ROUTE,
-        apply_generic_web_preview_verification,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebPreviewVerificationEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="apply_generic_web_preview_verification",
-        summary="Write generic web preview verification",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
+    register_generic_web_write_routes(
+        app,
+        dependencies=generic_web_write_route_dependencies,
+        handlers=generic_web_write_route_handlers,
     )
 
     app.add_api_route(
@@ -17334,62 +15262,10 @@ def create_launchplane_fastapi_app(
         },
     )
 
-    app.add_api_route(
-        _GENERIC_WEB_ROLLBACK_PLAN_ROUTE,
-        write_generic_web_rollback_plan,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebRollbackPlanEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="write_generic_web_rollback_plan",
-        summary="Plan generic web prod rollback",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
-    )
-
-    app.add_api_route(
-        _GENERIC_WEB_ROLLBACK_ROUTE,
-        write_generic_web_rollback,
-        methods=["POST"],
-        status_code=202,
-        response_model=AcceptedEvidenceResponse,
-        response_model_exclude_none=True,
-        openapi_extra={
-            "requestBody": {
-                "required": True,
-                "content": {
-                    "application/json": {
-                        "schema": _openapi_model_schema(GenericWebRollbackEnvelope)
-                    }
-                },
-            }
-        },
-        operation_id="write_generic_web_rollback",
-        summary="Execute generic web prod rollback",
-        responses={
-            400: {"model": LaunchplaneErrorResponse},
-            401: {"model": LaunchplaneErrorResponse},
-            403: {"model": LaunchplaneErrorResponse},
-            404: {"model": LaunchplaneErrorResponse},
-            409: {"model": LaunchplaneErrorResponse},
-            503: {"model": LaunchplaneErrorResponse},
-        },
+    register_generic_web_rollback_write_routes(
+        app,
+        dependencies=generic_web_write_route_dependencies,
+        handlers=generic_web_write_route_handlers,
     )
 
     app.add_api_route(
