+++ control_plane/http_routes/__init__.py
@@ -13,6 +13,13 @@ from control_plane.http_routes.evidence import (
     PromotionEvidenceRequest,
     register_evidence_write_routes,
 )
+from control_plane.http_routes.generic_web import (
+    GenericWebWriteRouteDependencies,
+    GenericWebWriteRouteHandlers,
+    build_generic_web_write_route_handlers,
+    register_generic_web_rollback_write_routes,
+    register_generic_web_write_routes,
+)
 from control_plane.http_routes.every_code import (
     register_every_code_feedback_read_routes,
     register_every_code_notification_attempt_read_routes,
@@ -26,6 +33,7 @@ from control_plane.http_routes.mutation_support import (
     accepted_evidence_response,
     idempotency_capable_store,
     idempotency_scope,
+    provider_operation_response_payload,
     replay_idempotent_response,
     request_fingerprint,
 )
@@ -66,14 +74,18 @@ __all__ = (
     "DriverReadRouteDependencies",
     "EVIDENCE_INGRESS_ROUTES",
     "EvidenceWriteRouteDependencies",
+    "GenericWebWriteRouteDependencies",
+    "GenericWebWriteRouteHandlers",
     "ProductReadRouteDependencies",
     "PromotionEvidenceRequest",
     "ReadRouteDependencies",
     "WorkGraphReadRouteDependencies",
     "accepted_evidence_response",
+    "build_generic_web_write_route_handlers",
     "idempotency_capable_store",
     "idempotency_scope",
     "product_profile_context_cutover_contexts_allowed",
+    "provider_operation_response_payload",
     "register_agent_context_read_routes",
     "register_deployment_promotion_read_routes",
     "register_dokploy_target_inspect_read_routes",
@@ -83,6 +95,8 @@ __all__ = (
     "register_every_code_notification_attempt_read_routes",
     "register_every_code_preview_gate_read_routes",
     "register_every_code_work_request_read_routes",
+    "register_generic_web_rollback_write_routes",
+    "register_generic_web_write_routes",
     "register_ingress_read_routes",
     "register_inventory_operation_read_routes",
     "register_managed_secret_read_routes",
