+++ tests/test_http_write_route_registrars.py
@@ -64,6 +64,121 @@ class FastApiWriteRouteRegistrarTests(unittest.TestCase):
         start = route_keys.index(expected[0])
         self.assertEqual(route_keys[start : start + len(expected)], expected)
 
+    def test_generic_web_write_routes_preserve_contracts_and_ownership(self) -> None:
+        expected_routes = (
+            (
+                "/v1/drivers/generic-web/preview-desired-state",
+                "apply_generic_web_preview_desired_state",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/preview-inventory",
+                "apply_generic_web_preview_inventory",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/preview-readiness",
+                "apply_generic_web_preview_readiness",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/preview-refresh",
+                "apply_generic_web_preview_refresh",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/preview-destroy",
+                "apply_generic_web_preview_destroy",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/deploy",
+                "apply_generic_web_deploy",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/prod-promotion",
+                "apply_generic_web_prod_promotion",
+                "GenericWebProdPromotionResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/prod-promotion-workflow",
+                "dispatch_generic_web_prod_promotion_workflow",
+                "GenericWebPromotionWorkflowResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/stable-verification",
+                "apply_generic_web_stable_verification",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/preview-verification",
+                "apply_generic_web_preview_verification",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/prod-rollback-plan",
+                "write_generic_web_rollback_plan",
+                "AcceptedEvidenceResponse",
+            ),
+            (
+                "/v1/drivers/generic-web/prod-rollback",
+                "write_generic_web_rollback",
+                "AcceptedEvidenceResponse",
+            ),
+        )
+        expected_paths = {path for path, _, _ in expected_routes}
+        extracted_routes = [
+            route
+            for route in self.api_routes
+            if route.methods == {"POST"} and route.path in expected_paths
+        ]
+
+        self.assertEqual(
+            [(route.path, route.name) for route in extracted_routes],
+            [(path, operation_id) for path, operation_id, _ in expected_routes],
+        )
+        for route, (_, operation_id, response_model) in zip(
+            extracted_routes, expected_routes, strict=True
+        ):
+            self.assertEqual(route.operation_id, operation_id)
+            self.assertEqual(route.response_model.__name__, response_model)
+            self.assertEqual(route.endpoint.__module__, "control_plane.http_routes.generic_web")
+
+    def test_generic_web_write_routes_preserve_interleaved_route_order(self) -> None:
+        route_keys = [(next(iter(route.methods or set())), route.path) for route in self.api_routes]
+        expected_primary = [
+            ("POST", "/v1/previews/lifecycle-sweep"),
+            ("POST", "/v1/drivers/generic-web/preview-desired-state"),
+            ("POST", "/v1/drivers/generic-web/preview-inventory"),
+            ("POST", "/v1/drivers/generic-web/preview-readiness"),
+            ("POST", "/v1/drivers/generic-web/preview-refresh"),
+            ("POST", "/v1/drivers/generic-web/preview-destroy"),
+            ("POST", "/v1/drivers/generic-web/deploy"),
+            ("POST", "/v1/drivers/generic-web/prod-promotion"),
+            ("POST", "/v1/drivers/generic-web/prod-promotion-workflow"),
+            ("POST", "/v1/drivers/generic-web/stable-verification"),
+            ("POST", "/v1/drivers/generic-web/preview-verification"),
+            ("POST", "/v1/drivers/verireel/testing-deploy"),
+        ]
+        expected_rollback = [
+            ("POST", "/v1/drivers/odoo/prod-backup-gate"),
+            ("POST", "/v1/drivers/generic-web/prod-rollback-plan"),
+            ("POST", "/v1/drivers/generic-web/prod-rollback"),
+            ("POST", "/v1/drivers/odoo/prod-rollback"),
+        ]
+
+        primary_start = route_keys.index(expected_primary[0])
+        rollback_start = route_keys.index(expected_rollback[0])
+        self.assertEqual(
+            route_keys[primary_start : primary_start + len(expected_primary)],
+            expected_primary,
+        )
+        self.assertEqual(
+            route_keys[rollback_start : rollback_start + len(expected_rollback)],
+            expected_rollback,
+        )
+
 
 if __name__ == "__main__":
     unittest.main()
